Heading Arrow Update
The previous way of getting orientation was pulling from a gps heading (I think) so relied on GPS for cardinal direction. The vivoactive 4s has a built in compass. Swapped it over to the compass sensor.
Note: from testing on my watch, the value given out is in radians with 0 rads = North. For orientation of the direction arrow I shifted it by Pi/2 to orient North to the top of the watch. Assuming it should show direction to North now.
Example Video:
https://photos.app.goo.gl/7jPmxkS24u7g7FXQA