Rotation Tweaks.
Fuck this shit.

I have no idea what I am missing but I cannot get the right angle for these test points and to render them correctly on the screen. More tweaking of things and finding the right rotation values and other shit. Also, Math.atan and Math.atan2 give two sepearte values... wtf? I'm guessing because the signs on the values are processed differently but shoudln't it just be the same? Testing and hand calculation will be needed.